UzayManga: Adding as a multisrc
